backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
openai==1.97.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5

